LayoutTests/accessibility/aria-combobox-hierarchy.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
<script src="../resources/accessibility-helper.js"></script>
</head>
<body id="body">
<div id="content">
  <div id="combobox" tabindex="0" role="combobox" aria-expanded="True">
    <div role="listbox">
      <div role="option">option 1</div>
      <div role="option">option 2</div>
    </div>
  </div>
</div>
<p id="description"></p>
<pre id="tree"></pre>
<div id="console"></div>
<script>
    description("This verifies the accessibility tree of ARIA comboboxes.");
    if (window.accessibilityController) {
        var axCombobox = accessibilityController.accessibleElementById("combobox");
        var stopElement = accessibilityController.accessibleElementById("stopElement");
        dumpAccessibilityTree(axCombobox, stopElement, 0);
        document.getElementById("content").style.visibility = "hidden";
    }
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
