LayoutTests/platform/glib/accessibility/aria-combobox-hierarchy-expected.txt

This verifies the accessibility tree of ARIA comboboxes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


AXRole: AXComboBox AXValue: option 1<\n>option 2<\n>
    AXRole: AXListBox
        AXRole: AXListItem AXValue: option 1
        AXRole: AXListItem AXValue: option 2
PASS successfullyParsed is true

TEST COMPLETE
